requirements.txt
# Requerimientos para el modelo de Gestos/Emociones:
streamlit
opencv-python
mediapipe
streamlit-webrtc
numpy

# Requerimientos para el semaforo electrónico:
pyserial